Launch local tgrade instance with DSO code set
