+++ scripts/tgrade/contracts/download_releases.sh
@@ -1,0 +1,23 @@
+#!/bin/bash
+set -o errexit -o nounset -o pipefail
+command -v shellcheck > /dev/null && shellcheck "$0"
+
+if [ $# -ne 1 ]; then
+  echo "Usage: ./download_releases.sh RELEASE_TAG"
+  exit 1
+fi
+
+tag="$1"
+# load token from OS keychain when not set via ENV
+GITHUB_API_TOKEN=${GITHUB_API_TOKEN:-"$(security find-generic-password -a "$USER" -s "github_api_key" -w)"}
+
+rm -f version.txt
+for contract in tgrade_dso; do
+  list_asset_url="https://api.github.com/repos/confio/tgrade-contracts/releases/tags/${tag}"
+  # get url for artifact with name==$artifact
+  asset_url=$(curl -H "Accept: application/vnd.github.v3+json" -H "Authorization: token $GITHUB_API_TOKEN" "${list_asset_url}" | jq -r ".assets[] | select(.name==\"${contract}.wasm\") | .url")
+  # download the artifact
+  curl -LJO -H 'Accept: application/octet-stream' -H "Authorization: token $GITHUB_API_TOKEN" "$asset_url"
+done
+
+echo "$tag" >version.txt
