+++ scripts/tgrade/contracts/version.txt
@@ -1,0 +1,1 @@
+v0.1.2
